typescript/rocketseat/back-solid-nodejs/src/usecases/seach-gyms/seach-gyms-usecase.ts
import { Gym } from '@prisma/client'
import { IGymsRepository } from '@/repositories/IGyms-repository'

interface SearchGymUseCaseRequest {
  query: string
  page: number
}

interface SearchGymUseCaseResponse {
  gyms: Gym[]
}

export class SearchGymUseCase {
  constructor(private gymRepository: IGymsRepository) {}

  async execute({
    query,
    page,
  }: SearchGymUseCaseRequest): Promise<SearchGymUseCaseResponse> {
    const gyms = await this.gymRepository.searchMany(query, page)

    return {
      gyms,
    }
  }
}
